#550 - server: further improvements
