+++ src/runtime/local/main.rs
@@ -83,7 +83,7 @@ fn options(mut argv: Vec<String>) -> Option<Vec<OptDef>> {
 }
 
 fn usage() {
-    println!("runtime: {}: [-h?pvcelq] [file...]", Mu::VERSION);
+    println!("mu-local: {}: [-h?pvcelq] [file...]", Mu::VERSION);
     println!("?: usage message");
     println!("h: usage message");
     println!("c: [name:value,...]");
