+++ src/runtime/server/server.rs
@@ -1,0 +1,161 @@
+//  SPDX-FileCopyrightText: Copyright 2022 James M. Putnam (putnamjm.design@gmail.com)
+//  SPDX-License-Identifier: MIT
+
+//! runtime loader/listener
+use {
+    crossbeam::channel,
+    futures::{
+        future::BoxFuture,
+        task::{self, ArcWake},
+    },
+    std::{
+        cell::RefCell,
+        future::Future,
+        pin::Pin,
+        sync::{Arc, Mutex},
+        task::{Context, Poll, Waker},
+        thread,
+        time::{Duration, Instant},
+    },
+};
+
+//
+/// server
+//
+pub struct Server {
+    scheduled: channel::Receiver<Arc<Task>>,
+    sender: channel::Sender<Arc<Task>>,
+}
+
+impl Default for Server {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Server {
+    pub fn new() -> Self {
+        let (sender, scheduled) = channel::unbounded();
+
+        Self { scheduled, sender }
+    }
+
+    pub fn spawn<F>(&self, future: F)
+    where
+        F: Future<Output = ()> + Send + 'static,
+    {
+        Task::spawn(future, &self.sender);
+    }
+
+    pub fn run(&self) {
+        CURRENT.with(|cell| {
+            *cell.borrow_mut() = Some(self.sender.clone());
+        });
+
+        while let Ok(task) = self.scheduled.recv() {
+            task.poll();
+        }
+    }
+}
+
+pub fn spawn<F>(future: F)
+where
+    F: Future<Output = ()> + Send + 'static,
+{
+    CURRENT.with(|cell| {
+        let borrow = cell.borrow();
+        let sender = borrow.as_ref().unwrap();
+
+        Task::spawn(future, sender);
+    });
+}
+
+pub async fn delay(dur: Duration) {
+    struct Delay {
+        when: Instant,
+        waker: Option<Arc<Mutex<Waker>>>,
+    }
+
+    impl Future for Delay {
+        type Output = ();
+
+        fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<()> {
+            if let Some(waker) = &self.waker {
+                let mut waker = waker.lock().unwrap();
+
+                if !waker.will_wake(cx.waker()) {
+                    *waker = cx.waker().clone();
+                }
+            } else {
+                let when = self.when;
+                let waker = Arc::new(Mutex::new(cx.waker().clone()));
+                self.waker = Some(waker.clone());
+
+                thread::spawn(move || {
+                    let now = Instant::now();
+
+                    if now < when {
+                        thread::sleep(when - now);
+                    }
+
+                    let waker = waker.lock().unwrap();
+                    waker.wake_by_ref();
+                });
+            }
+
+            if Instant::now() >= self.when {
+                Poll::Ready(())
+            } else {
+                Poll::Pending
+            }
+        }
+    }
+
+    let future = Delay {
+        when: Instant::now() + dur,
+        waker: None,
+    };
+
+    future.await;
+}
+
+//
+// task engine
+//
+thread_local! {
+    static CURRENT: RefCell<Option<channel::Sender<Arc<Task>>>> =
+        RefCell::new(None);
+}
+
+struct Task {
+    future: Mutex<BoxFuture<'static, ()>>,
+    executor: channel::Sender<Arc<Task>>,
+}
+
+impl Task {
+    fn spawn<F>(future: F, sender: &channel::Sender<Arc<Task>>)
+    where
+        F: Future<Output = ()> + Send + 'static,
+    {
+        let task = Arc::new(Task {
+            future: Mutex::new(Box::pin(future)),
+            executor: sender.clone(),
+        });
+
+        let _ = sender.send(task);
+    }
+
+    fn poll(self: Arc<Self>) {
+        let waker = task::waker(self.clone());
+        let mut cx = Context::from_waker(&waker);
+        let mut future = self.future.try_lock().unwrap();
+
+        let _ = future.as_mut().poll(&mut cx);
+    }
+}
+
+impl ArcWake for Task {
+    fn wake_by_ref(arc_self: &Arc<Self>) {
+        let _ = arc_self.executor.send(arc_self.clone());
+    }
+}
