+++ tests/test-ns.py
@@ -9,12 +9,12 @@ with open(base + '/' + ns + '/tests') as f: group_list = f.readlines()
 
 def runtest(line, group, test, expected):
     if ns == 'mu':
-        proc = subprocess.Popen(['../dist/local', '-p', '-e' + test],\
+        proc = subprocess.Popen(['../dist/mu-local', '-p', '-e' + test],\
                                 stdout=subprocess.PIPE,\
                                 stderr=subprocess.PIPE)
 
     if ns == 'core':
-        proc = subprocess.Popen(['../dist/local',
+        proc = subprocess.Popen(['../dist/mu-local',
                                  '-l../dist/core.l',
                                  '-q (core:%init-core-ns)',
                                  '-p',
